edit q4
